[Patch] A piece of code to test your config optimization
[Do not merge, this is not meant to be final code]
I wondered wether my config was optimal on training, so I added a timing over the batch generator to measure what was the slowest between image preprocessing and model training.

As the batch generator is the pivot between the both, I added timing on "put" in the queue and "get" in the queue. This seems interesting as on my config (Core i5, GFX 1070) the "get" takes more time than the put, which indicates that the put is not fast enough, so my CPU seems to be too slow...

Feel free to test it if you want and share your result. Maybe I'm wrong on my assumptions...