Several SQlite3 based optimizations: - DB stores filenames, status of extraction and conversion, and if marks frames where faces where detected.
This allows to skip already processed frames based on the DB data, instead of reading and comparing input and output directories. Works for Extract and Convert using the --skip-existing argument.

For conversion, as the DB stores frames where faces were found, it's possible to only run conversion on those frames using the --only-frames-with-faces argument.

Added a new script named db that allows to flush the Database when new data is going to be used.